requirements.txt
flask
numpy
open-clip-torch
torch
pillow
scikit-learn
matplotlib
pandas